Rename project feature works
