+++ packages/client/src/components/app/AppRoutes.tsx
@@ -3,7 +3,6 @@ import { Route, Switch } from "react-router-dom";
 import Project from "../project";
 import About from "../about";
 import Dataset from "../dataset";
-import Projects from "../projects";
 
 export default function AppRoutes(): ReactElement {
   return (
@@ -17,7 +16,6 @@ export default function AppRoutes(): ReactElement {
           return <Dataset id={match.params.id} />;
         }}
       />
-      <Route path="/projects" exact component={Projects} />
       <Route
         exact
         path="/projects/:id"
