--- packages/client/src/components/datasets/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Datasets";
