+++ packages/client/src/components/project/DeleteDatasetDialog.tsx
@@ -1,54 +1,18 @@
-import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  IconButton,
-  Menu,
-  MenuItem,
-  TextField
-} from "@material-ui/core";
-import { GetProjectList_projects } from "./__generated__/GetProjectList";
+import { Button, Dialog, DialogActions, DialogContent, DialogTitle, TextField } from "@material-ui/core";
 import { useApolloClient, useMutation } from "@apollo/client";
-import MoreVertIcon from "@material-ui/icons/MoreVert";
 import React, { ReactElement } from "react";
 import gql from "graphql-tag";
+import { GetProjectData_project } from "./__generated__/GetProjectData";
 
 const DELETE_PROJECT_MUTATION = gql`
-  mutation DeleteProject($id: ID!) {
+  mutation DeleteProjectM($id: ID!) {
     deleteProject(id: $id) {
       id
     }
   }
 `;
 
-export default function ProjectMenu(props: { project: GetProjectList_projects }): ReactElement {
-  const [anchorEl, setAnchorEl] = React.useState<Element | null>(null);
-
-  const handleClick = (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setAnchorEl(null);
-  };
-
-  return (
-    <>
-      <IconButton onClick={handleClick}>
-        <MoreVertIcon />
-      </IconButton>
-      <Menu id="simple-menu" anchorEl={anchorEl} keepMounted open={Boolean(anchorEl)} onClose={handleClose}>
-        <MenuItem onClick={handleClose}>
-          <RemoveProjectDialogButton project={props.project} />
-        </MenuItem>
-      </Menu>
-    </>
-  );
-}
-
-function RemoveProjectDialogButton(props: { project: GetProjectList_projects }): ReactElement {
+export default function DeleteProjectDialogButton(props: { project: GetProjectData_project }): ReactElement {
   const [open, setOpen] = React.useState<boolean>(false);
   const [confirmation, setConfirmation] = React.useState<string>("");
   const [deleteProject] = useMutation(DELETE_PROJECT_MUTATION);
@@ -71,7 +35,7 @@ function RemoveProjectDialogButton(props: { project: GetProjectList_projects }):
 
   return (
     <>
-      <Button onClick={handleClickOpen}>Delete Project</Button>
+      <Button onClick={handleClickOpen}>Delete</Button>
       <Dialog onClose={handleClose} open={open}>
         <DialogTitle>Confirm Project Deletion</DialogTitle>
         <DialogContent dividers>
