--- packages/client/src/components/projects/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./Projects";
